Room2 Lever and Ramp
Working script for pushing the lever which moves ramp to position
