+++ Games studio/Assets/Scripts/LeverScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 90a7698d4d6917b4882313f1b7b2507b
+guid: 6a93ca7619c4da144b4d7a292485387b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
